Replace crate seccomp-sys with libseccomp-sys
The crate seccomp-sys is licensed under the LGPLv2.1, which is
problematic because code from crates is not dynamically linked, but
compiled and linked into the binary.

Replace it with libseccomp-sys, which also provides the bindings we
need (in fact, we only need to change the reference to the crate), and
it's licensed under MIT or Apache-2.0.

As a nice bonus, the bindings in libseccomp-sys are newer, meaning we
can now support the "log" action for seccomp for real. We'll do that
in the next commit.

Signed-off-by: Sergio Lopez <slp@redhat.com>