Add support the "log" action for seccomp
In 5f63ba4 ("Replace Firecraker's seccomp with seccomp-sys") we
replaced Firecracker's seccomp with the "seccomp-sys" crate. With this
change, we losed the ability to use the "log" action for the "seccomp"
argument (and we forgot to update the argument's documentation, BTW
;-).

In the previous commit, we replaced "seccomp-sys" with
"libseccomp-sys", which provides newer bindings including the ones we
need for re-enabling the "log" action. Do it so.

Signed-off-by: Sergio Lopez <slp@redhat.com>