Set signal handlers, required for an "init" process
The first process created in a new PID namespace (pid_namespaces(7))
is the "init" process, for that namespace. Only signals for which the
"init" process has established a  signal handler can be sent to it.
To run  virtiofsd with '--sandbox namespace', we need to set handlers
to end the child process if the parent dies. Also, in case the
PID namespace is set by an external tool. Fixes #23

Signed-off-by: German Maglione <gmaglione@redhat.com>