virtiofsd: Update fuse and bring it to latest minor version 36
I see that in rust we are bringing in fuse changes but not updating
minor version. This is kind of odd. We should bring in all relevant
fuse header changes upto a certain versin and then update
KERNEL_MINOR_VERSION accordingly. This gives a clear picture of
upto what minor version server understands and gives correct
information back to client (kernel).

So update fuse to bring it to latest minor version 36. There should
not be any functional changes as I am not opting in for any
additional new features. Only significant change is that it can
now look at INIT_EXT flag and be able to parse both extended as
well as old fuse_init_in struct.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>