Check whether the root's file handle is usable
Add a check to `PassthroughFs::new()` to verify whether file handles are
actually usable.  To do this, the new `check_working_file_handles()`
method will open the root directory (i.e. the shared directory), turn
the FD into a file handle, and open that file handle.  It is possible
that the last step fails (e.g. because we believe we have the
`DAC_READ_SEARCH` capability, but actually we only have it in our user
namespace, which is insufficient), and then we need to detect this
before we actually begin start using file handles.

In "fallback" mode, `check_working_file_handles()` will emit a warning
and switch to "never" mode.  In "mandatory" mode, we will error out and
consequentially exit the whole process.

Similarly, if no file handle could be generated for this root node,
disable file handles altogether, too, because we cannot be certain that
they are usable.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>