Add PassthroughFs::make_file_handle_openable()
We already have the same piece of code in two places, and the next patch
is going to add a third one.  Add a small helper function to turn a file
handle into an openable file handle in the context of PassthroughFs
(i.e. PassthroughFs.mount_fds and PassthroughFs.proc_self_fd).

Signed-off-by: Hanna Reitz <hreitz@redhat.com>