Add "prefer" alias to --inode-file-handles
"fallback" as a mode name may be misleading, because it is not
immediately clear whether the file handle is the fallback, or something
else.  (It is something else.)

Add a "prefer" alias that makes it clearer that in this mode we prefer
to generate file handles, but which implies that we can do something
else if that does not work out.

Document this value as the preferred value, mentioning "fallback" only
as an alias.  (Use this opportunity to also add "mandatory" to the
shortlist of supported --inode-file-handles values, which we forgot in
commit c329bfda7bb66b59ac4ee77b7131c309f02f0e44.)

Fixes: #20
Suggested-by: Sebastian Hasler <sebastian.hasler@stuvus.uni-stuttgart.de>
Signed-off-by: Hanna Reitz <hreitz@redhat.com>