Set the number of written bytes for used descs
As the Linux driver ignores the "len" field of used descriptors, we
didn't bother to set it properly when return those descriptors to the
queue.

The problem is that other implementations (at least, the Windows one),
do care about it, so let's do the right thing and set it properly.

Signed-off-by: Sergio Lopez <slp@redhat.com>