Add an option to enable/disable posix acls
Add virtiofsd option '--posix-acl' to let users enable posix ACLs
support, and the compatibility options '-o posix_acl/no_posix_acl'.
Fuse has an option FUSE_POSIX_ACL which needs to be opted in by fuse
server to enable posix ACLs. If the client don't support posix ACLs,
virtiofsd will log an error and replies EPROTO to the client.
By default, it's disabled due to performance concerns with
'--cache=never'.

So, if posix ACLs are enabled, opt in for FUSE_DONT_MASK. virtiofsd
will set umask of the thread doing file creation: create, mknod and
mkdir, excluding symlink which doesn't receive umask as a parameter.
And host kernel should use that umask if parent directory does not have
default ACLs, otherwise umask does not take effect.

Also, when posix access ACLs are set on a file, it can lead to
adjusting file permissions (mode) as well. If caller does not have
CAP_FSETID and it also does not have membership of owner group,
this will lead to clearing SGID bit in mode.
Fuse expects the file server to take care of clearing SGID,
the client will set the flag FUSE_SETXATTR_ACL_KILL_SGID in setxattr
request. In that case, the server should kill SGID.
The host kernel does not clear the SGID bit if virtiofsd is
running as root and has CAP_FSETID. So, virtiofsd need to switch to gid
of caller in guest, and also drop CAP_FSETID (if caller did not have it
to begin with).

Signed-off-by: German Maglione <gmaglione@redhat.com>