Move 'umask(0)' call to the main thread
We need to set umask to 0, so we don't mask the creation mode,
the guest kernel already does this. Currently, 'umask(0)' is called
in 'PassthroughFs::init()', thus it's a child thread that sets umask
to 0. This will be a problem when we stop sharing the file system
attributes (calling 'unshare(CLONE_FS)' in the next commit), because
changing umask in a child thread will not affect another thread.
Therefore, this commit moves the call to 'umask(0)' to
'PassthroughFs::new()' to be called in the main thread.

Signed-off-by: German Maglione <gmaglione@redhat.com>