Increment the rlimit_nofile for all sandbox modes
Currently, we only increment the rlimit_nofile with the user-supplied
value within sandbox modes 'chroot' and 'namespace'. However, in the
'namespace' mode, we don't honor '--rlimit-nofile=0' and we increase
the value to the maximum, as in the default case.

This commit try to increase the rlimit_nofile for all modes,
honoring '--rlimit-nofile=0'. Additionally, it leaves the limit
unchanged if the limit is greater than 1M. Also, if everything fails
the hard limit is used.

Signed-off-by: German Maglione <gmaglione@redhat.com>