Do not drop supplemental groups for unprivileged user namespace
Right now we call setgroups() and drop all supplementary groups for all
the sandboxing modes. This does not work for the case of "--sandbox=Namespace"
when launched by unprivileged user as we don't have permission to call
setgroups()  (neither outside namespace nor inside namespace).

When running unprivileged and not using newuidmap/newgidmap, we have ownership
of only unprivileged id and can map only that user. Previous patch added
the code to create 1:1 mapping of uid/gid. That means there is no arbitray
switching. All the files owned by user show up as same uid inside guest.
Andything not owned by uid/gid showsup as "nobody".

Given there is no arbitrary switching of uid/gid and everything in the
fs is owned by one uid/gid (or nobody), I think a particular
issue which required clearing setgroups() does not arise in this
configuration. So it might not be a necessacity to drop membership
of all supplementary groups.

So with "--sandbox=Namespace" launched by unprivileged user, do not
drop supplementary groups (setgroups()). This is get this mode working
again. 

If this becomes a problem ever again, then we will have to look at
using newuidmap/newgidmap so that we have capability to call setgroups()
inside user namespace and then we can call setgroups() and drop all
the supplementary groups. 

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>