Fix daemon startup with new glibc versions (>=2.35)
When running on systems that ship with the most recent release of glibc
(2.35), we noticed that virtiofsd would fail when running as a regular
user. This was down to one of viriofsd's child processes being killed by
seccomp for attempting an rseq. So, add rseq to allowed syscalls, and
bump libc bindings to the version that introduces SYS_rseq.

Signed-off-by: Anca Iordache <anca.iordache@docker.com>
Signed-off-by: Frédéric Dalleau <frederic.dalleau@docker.com>
Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>