Allow only one daemon to bind the socket
Currently, more than one daemon can start using the same unix socket
name. This try to get a lock file and storing the pid in that file,
before deleting the socket file.

The implementation of write_pid_file prevents a race condition if the
pid file is deleted _before_ being closed. We don't delete the pid
file, for now, because it would not work when changing the root dir
with --sandbox=chroot/namespace, this will be fixed in the future.

Signed-off-by: German Maglione <gmaglione@redhat.com>