Fix st_mode of . and .. directories in readdirplus
Generally, st_mode (as a field of struct stat) is made of S_IFDIR,
S_IFREG, etc., but d_type (as a field of struct dirent) is made of
DT_DIR, DT_REG, etc. But these definitions are differ up to a 12 bit
shift.
It leads to incorrect attributes of . and .. in Windows guest, so fix it.

Signed-off-by: Viktor Prutyanov <viktor.prutyanov@redhat.com>