Do not create tmpdir and bind mount /proc/self/fd on /proc instead
virtiofsd needs access to /proc/self/fd but we also want to make
sure that if virtiofsd is hacked, it is not able to traverse up the
directory tree and be able to access /proc/ etc.

Currently virtiofsd creates a tmpdir and bind mounts /proc/self/fd/
there and then opens an O_PATH fd for tmpdir. And then it deletes
the tmpdir.

C version of virtiofsd bind mounts /proc/self/fd on /proc/ and then
opens an O_PATH fd on /proc.

Creating a separate dir needs to find a temporary space and SELinux
policy might not allow for writing to some tmp space directly. Instead
runtime needs to setup some space and label it properly so that virtiofsd
can write to it.

So going away from the need of creating tmpdir will be good as long
as one can not traverse into /proc/self or /proc. I think it is
not possible with bind mounting /proc/self/fd on /proc. So proposing
this patch. 

BTW, I noticed that in sandbox=chroot path we don't use any tricks
to block compromised virtiofsd to get to /proc. Stefan mentioned that
reason being that we have not setup mount namespace. And if we bind
mount /proc/self/fd onto /proc, it will break other processes using
/proc. So may be tempdir trick can be used for chroot() case as long
as we figure out a place where tempdir can be created and SELinux
allows it.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>